% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ljosa2013.R
\docType{data}
\name{ljosa2013}
\alias{ljosa2013}
\title{Image metadata and URLs for Ljosa et al., J. Biomol. Screening, 2013.}
\format{An object of class \code{tbl_df} (inherits from \code{data.frame}) with 2528 rows and 12 columns.}
\source{
\url{https://data.broadinstitute.org/bbbc/BBBC021/}
}
\usage{
ljosa2013
}
\description{
Image metadata and URLs for Ljosa et al., J. Biomol. Screening, 2013.
}
\keyword{datasets}
