% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_max_intensity.R
\name{sample_max_intensity}
\alias{sample_max_intensity}
\title{Robust max intensity}
\usage{
sample_max_intensity(df, is_url = TRUE)
}
\arguments{
\item{df}{...}

\item{is_url}{...}
}
\value{
robust max intensities per channel
}
\description{
Robust max intensity
}

