% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_images.R
\name{save_images}
\alias{save_images}
\title{Save images}
\usage{
save_images(df, per_row = 3, max_intensity = NULL, type = "jpg",
  quality = 50, output_dir = ".", is_url = TRUE)
}
\arguments{
\item{df}{...}

\item{per_row}{...}

\item{max_intensity}{...}

\item{type}{...}

\item{quality}{...}

\item{output_dir}{...}

\item{is_url}{...}
}
\value{
list of file names
}
\description{
Save images
}
